- 배송요청 루틴 작업중 - 배송 물품정보 서버값으로 변경

+++ app/src/main/java/sky/kr/co/newtogetusa/data/remote/dto/delivery/DeliveryResponse.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class DeliveryResponse(
     @SerializedName("delivery_id")
-    val deliveryId: Int
+    val deliveryId: Long
 )
